SectionFilling: Warn if a bear is not found
When loading bears, we do not check if all the bears given were
found. This checks the list of bears that we have with the list of
bears that were found. Then we print warnings to tell the user that
the specified bear was not found.

Fixes https://github.com/coala-analyzer/coala/issues/909